feat(design): AgentChatInput's layout carves into compound parts (#8676)
